Добавлен роутинг в приложение. Чат вынесен в отдельную вкладку и добавлена заготовка страницы профиля.
