+++ Docs/task-03.md
@@ -1,6 +1,6 @@
 ﻿*[Назад](./../README.md)*  
   
-### Задание для урока №2  
+### Задание для урока №3  
   
 - [X] Установить material-ui. Переделать форму с использованием компонентов из material-ui.  
 - [X] Добавить автофокус на текстовое поле при открытии страницы и после отправки сообщения.
