+++ Docs/task-04.md
@@ -1,10 +1,10 @@
 ﻿*[Назад](./../README.md)*  
   
-### Задание для урока №2  
+### Задание для урока №4  
   
-- [ ] Установить react-router-dom. Добавить домашнюю страницу по адресу "/" со 
+- [X] Установить react-router-dom. Добавить домашнюю страницу по адресу "/" со 
 списком ссылок на страницу чатов и страницу профиля.  
-- [ ] Добавить страницу профиля (пока не несет никакой функциональности, 
+- [X] Добавить страницу профиля (пока не несет никакой функциональности, 
 можно сделать ее пустой).  
 - [ ] Настроить разделение приложения на чаты с помощью роутера (использовать параметры url). 
 Приложение должно корректно работать, если пользователь вводит идентификатор несуществующего
