+++ lesson-04/src/Profile/index.js
@@ -1,0 +1,6 @@
+const Profile = () => {
+    return <div>Страница с описанием профиля пользователя</div>;
+};
+  
+export default Profile;
+
